Refactored hosting using nginx
This change implements a more systematic approach to hosting web
applications. There's a new package called `selfhost` which provides
three interfaces: domains (currently only clearnet, onion in the future),
applications and gateway. Gateway is currently hard-coded to nginx and
an improvement will need another (smaller) refactor, but the foundation
is there. Finally clearnet provides an interface for specifying TLS
certificates, which can be automated using certbot.

This uses the new features of debcrafter in order to ensure proper
cleanup.

Closes #47
Closes #44