Fixed generation of apps and domains
Due to a bug in mustache, generation broke when there was more than one
app or more than one domain. This change generates the configuration
based on individual files. This is even better because the system
can be more robust against misconfiguration. That being said, it's still
not as robust as I'd like to.

Aside from that, the new generator script checks domain length and
adjusts hash bucket size accordingly.

Closes #54
Closes #56