Improved dependency relationships
The dependencies between service and confext packages were inverted in
order to avoid restarting services, especially bitcoind which takes a
long time to restart. This speeds up the installation time considerably.

Further `bitcoin-pruned-mainnet` package was created which is the
default for `bitcoin-chain-mode` - default package and also allows to
specify the exact prune size (550 by default, of course).

Closes #57