Fixed missing user for ridetheln-system
When ridetheln-lnd-system-mainnet is installed, it adds ridetheln-system
into appropriate group to have access to lnd macaroon. However, the user
didn't exist at the time of installation due to changed order of
dependencies. As a quick fix, the order was changed back for this
instance and will need to be done properly later.