Fixed upgrade/reinstall in tor-hs-patch-config
When upgrading or reinstalling `tor-hs-patch-config`, it removed the
diversion in postrm, which is only executed after preinst of the
replacing package was executed. However, since preinst attmpted to
create the same diversion, it was left unchange, turning it into no-op.
The postrm then proceeded to remove it and the postinst in the next
version overwrote the file completely.

This was fixed by moving the removal logic to prerm, which puts the
system into the state it was in before `tor-hs-patch-config` was
installed. Thus the installation of the next package will work the same
as on a clean system.