Bitcoind cookie file moved to /var/run/
/var/run is the appropriate place to put runtime files such as the
cookie files. This change was planned before and was accidentally
considered needed to solve other issue, but that wasn't actually the
case. :)

However, once it's done, it's already useful because it solves the
security issue when bitcoind dies and malicious program could bind the
same port in order to get the password.

Thanks to this, #9 is not needed in case of bitcoind.