Robust reload-or-restart
`reload-or-restart` doesn't seem to be working according to the
documentation and does not start the service if it's not running. This
caused issues in case of `tor-hs-config-patch` and was fixed by explicit
check. This is racy, obviously, but probably the only thing we can do
and hopefully not a huge issue.