Fixed bitcoin-pruned-mainnet not being default
There was an issue reported by @lukechilds that caused bitcoin-mainnet
to depend on arbitrary chain mode by default instead of pruned one.

This change is a quick fix that hacks it by depending on dpkg, which
will always be installed. Improvements to dependency ordering are
planned in the future.

This also adds two more tests that prevent this from repeating again as
well as a test to make sure txindex is not turned on needlessly.