Implemented auto unlock for LND
This change adds a script and packages for automatically unlocking LND.
The way it works is it generates the seed and stores it in a file, which is
then used to init the wallet with a fixed password and unlock it later.

This does NOT decrease the security because if an attacker gains access
to the seed file or private data he can also gain the access to the
macaroon file, which LND doesn't protect. The whole design of LND
password is mostly counterproductive as discussed in
https://github.com/lightningnetwork/lnd/issues/899

The unlocker uses `wget` instead of `lncli` because `lncli` doesn't
allow non-interactive initialization of the wallet. This forcces us to
use other tool (``wget`), which is used for unlocking too, for
consistency.

This is currently implemented for mainnet only. Regtest will come soon.