Added a bash library for accessing LND
This commit adds a bash library that can be used to get the connection
information. It's considered a public API and the change also uses it in
`xlncli`, `auto_unlock.sh` and tests.

Additionally, this change also changes `xlncli` to defer to `lncli`
whenever *any* of the arguments changing LND access are present. This
is less confusing and avoids leaks of macaroons. It also makes sure to
not require macaroon path for `create` and `unlock`.