Attempted update of bitcoind to 0.21 - BROKEN!
**Warning: `bitcoind` 0.21 seems to be broken as it's sometimes stuck
during startup. Do NOT use this change!**

This change is commited into a separate branch for documentation/backup
purposes.

This updates bitcoind (and cli) to 0.21 and uses its new feature
`startupnotify` to improve robustness.
It calls `systemd-notify` to postpone starting of depending services and
also removes reindex marker file *after* startup so that crash/power
outage can not lead to not reindexing.

Would have closed issues #106 and #110.