Created package cadr-policy
This package manages settings common to our packages. Currently there's
only one setting - disabling unneeded services.

Unneeded service is one added by the package that also carries binaries
and with configuration that is largely unsuitable for our use, so that
we created our own. This default service is thus not needed unless the
user uses it for something else. Scripts in `cadr-policy` try to detect
such situation and do not disable the service if its package is manually
installed or depended on by package that is not ours.