Improve handling of FPM pool config
The PHP FPM pool configuration was previously copied into
/etc/php/$version/fpm/pool.d as a non-conffile which was not great as it
was neither managed by dpkg nor auto-generated. This moves the file to
/etc/nextcloud-server-system and puts symlink to it at previous
location. This way it can become a conffile without path changing being
an issue.

Additionally, instead of comparing versions which was error-prone, this
just checks for existence/non-existence of the links. It solves some
edge case bugs and is a more principled approach to idempotent scripts.