Print journal of service if installation fails
When a service fails to start the maintainerscript would only print a
help, not full log. This is annoying when testing in CI as there's no
way to do that and even in case of local tests it's just one more step.

This change adds a check that would print the log for service packages.
Although it won't show anything if the failure happened before starting,
this is OK, it's just debug info.